warnings for failed nodes in dynamic execution
